Move errors related to metadata
